fix: Break IPlatformSpecifics contents into individual files.
