+++ com.playeveryware.eos/Runtime/Core/IEOSCoroutineOwner.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b5c3a175a6ed8dc4da94dfd98dd9d3dd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
